仔细看了一下Multiboot规范，大致理解了调试函数（不过要是让我自己写恐怕写不出来）。因为printk函数的实现有问题所以停滞了很久。之前的printk函数的实现是参考的xv6源码，现在看来是我抄错了。接下来要仔细看以下那位兄弟的printk函数了。

+++ include/elf.h
@@ -29,7 +29,7 @@ typedef struct elf_symbol_t{
     uint32_t size;
     uint8_t info;
     uint8_t other;
-    uint8_t shndx;
+    uint16_t shndx;
 }__attribute__((packed)) elf_symbol_t;
 
 //information
